+++ src/android/TriggerReceiver.java
@@ -84,8 +84,11 @@ public class TriggerReceiver extends AbstractTriggerReceiver {
      */
     @Override
     public Notification buildNotification(Builder builder, Bundle bundle) {
-        return builder.setClickActivity(ClickReceiver.class).setClearReceiver(ClearReceiver.class).setExtras(bundle)
-                .build();
+      return builder
+      .setClickActivity(ClickReceiver.class)
+      .setClearReceiver(ClearReceiver.class)
+      .setExtras(bundle)
+      .build();
     }
 
 }