+++ src/android/notification/Builder.java
@@ -2,6 +2,7 @@
  * Apache 2.0 License
  *
  * Copyright (c) Sebastian Katzer 2017
+ * Contributor fquirin, Bhumin Bhandari, powowbox
  *
  * This file contains Original Code and/or Modifications of Original Code
  * as defined in and that are subject to the Apache License
@@ -35,16 +36,14 @@ import android.graphics.Color;
 import android.graphics.PorterDuff;
 import android.graphics.PorterDuffXfermode;
 import android.graphics.Rect;
-import android.graphics.RectF;
 import android.graphics.Paint;
 import android.graphics.Canvas;
 
 import java.util.List;
-import java.util.Random;
 
 import de.appplant.cordova.plugin.notification.action.Action;
+import de.appplant.cordova.plugin.notification.util.LaunchUtils;
 
-import static android.app.PendingIntent.FLAG_UPDATE_CURRENT;
 import static de.appplant.cordova.plugin.notification.Notification.EXTRA_UPDATE;
 
 /**
@@ -58,9 +57,6 @@ public final class Builder {
 
     // Notification options passed by JS
     private final Options options;
-
-    // To generate unique request codes
-    private final Random random = new Random();
 
     // Receiver to handle the clear event
     private Class<?> clearReceiver;
@@ -196,15 +192,14 @@ public final class Builder {
     void applyFullScreenIntent(NotificationCompat.Builder builder) {
         String pkgName  = context.getPackageName();
 
+        int notificationId  = options.getId();
         Intent intent = context
             .getPackageManager()
             .getLaunchIntentForPackage(pkgName)
-            .putExtra("launchNotificationId", options.getId());
+            .putExtra("launchNotificationId", notificationId);
 
-        int reqCode = random.nextInt();
-        // request code and flags not added for demo purposes
-        PendingIntent pendingIntent = PendingIntent.getActivity(context, reqCode, intent, FLAG_UPDATE_CURRENT);
-
+        PendingIntent pendingIntent =
+          LaunchUtils.getActivityPendingIntent(context, intent, notificationId);
         builder.setFullScreenIntent(pendingIntent, true);
     }
 
@@ -226,14 +221,14 @@ public final class Builder {
         final int color = Color.RED;
         final Paint paint = new Paint();
         final Rect rect = new Rect(0, 0, bitmap.getWidth(), bitmap.getHeight());
-        final RectF rectF = new RectF(rect);
+        //final RectF rectF = new RectF(rect);
 
         paint.setAntiAlias(true);
         canvas.drawARGB(0, 0, 0, 0);
         paint.setColor(color);
-        float cx = bitmap.getWidth() / 2;
-        float cy = bitmap.getHeight() / 2;
-        float radius = cx < cy ? cx : cy;
+        float cx = bitmap.getWidth() / 2.0f;
+        float cy = bitmap.getHeight() / 2.0f;
+        float radius = Math.min(cx, cy);
         canvas.drawCircle(cx, cy, radius, paint);
 
         paint.setXfermode(new PorterDuffXfermode(PorterDuff.Mode.SRC_IN));
@@ -388,19 +383,17 @@ public final class Builder {
         if (clearReceiver == null)
             return;
 
+        int notificationId = options.getId();
         Intent intent = new Intent(context, clearReceiver)
                 .setAction(options.getIdentifier())
-                .putExtra(Notification.EXTRA_ID, options.getId());
+                .putExtra(Notification.EXTRA_ID, notificationId);
 
         if (extras != null) {
             intent.putExtras(extras);
         }
 
-        int reqCode = random.nextInt();
-
-        PendingIntent deleteIntent = PendingIntent.getBroadcast(
-                context, reqCode, intent, FLAG_UPDATE_CURRENT);
-
+        PendingIntent deleteIntent =
+          LaunchUtils.getBroadcastPendingIntent(context, intent, notificationId);
         builder.setDeleteIntent(deleteIntent);
     }
 
@@ -415,8 +408,16 @@ public final class Builder {
         if (clickActivity == null)
             return;
 
+        Action[] actions = options.getActions();
+        if (actions != null && actions.length > 0 ) {
+          // if actions are defined, the user must click on button actions to launch the app.
+          // Don't make the notification clickable in this case
+          return;
+        }
+
+        int notificationId  =  options.getId();
         Intent intent = new Intent(context, clickActivity)
-                .putExtra(Notification.EXTRA_ID, options.getId())
+                .putExtra(Notification.EXTRA_ID, notificationId)
                 .putExtra(Action.EXTRA_ID, Action.CLICK_ACTION_ID)
                 .putExtra(Options.EXTRA_LAUNCH, options.isLaunchingApp())
                 .setFlags(Intent.FLAG_ACTIVITY_NO_HISTORY);
@@ -425,11 +426,8 @@ public final class Builder {
             intent.putExtras(extras);
         }
 
-        int reqCode = random.nextInt();
-
-        PendingIntent contentIntent = PendingIntent.getService(
-                context, reqCode, intent, FLAG_UPDATE_CURRENT);
-
+        PendingIntent contentIntent =
+          LaunchUtils.getTaskStackPendingIntent(context, intent, notificationId);
         builder.setContentIntent(contentIntent);
     }
 
@@ -465,8 +463,9 @@ public final class Builder {
      * @param action Notification action needing the PendingIntent
      */
     private PendingIntent getPendingIntentForAction (Action action) {
+        int notificationId =  options.getId();
         Intent intent = new Intent(context, clickActivity)
-                .putExtra(Notification.EXTRA_ID, options.getId())
+                .putExtra(Notification.EXTRA_ID, notificationId)
                 .putExtra(Action.EXTRA_ID, action.getId())
                 .putExtra(Options.EXTRA_LAUNCH, action.isLaunchingApp())
                 .setFlags(Intent.FLAG_ACTIVITY_NO_HISTORY);
@@ -475,10 +474,7 @@ public final class Builder {
             intent.putExtras(extras);
         }
 
-        int reqCode = random.nextInt();
-
-        return PendingIntent.getService(
-                context, reqCode, intent, FLAG_UPDATE_CURRENT);
+      return LaunchUtils.getTaskStackPendingIntent(context, intent, notificationId);
     }
 
     /**
@@ -487,7 +483,8 @@ public final class Builder {
      * @return true in case of an updated version.
      */
     private boolean isUpdate() {
-        return extras != null && extras.getBoolean(EXTRA_UPDATE, false);
+        return extras != null
+            && extras.getBoolean(EXTRA_UPDATE, false);
     }
 
     /**