+++ src/android/notification/receiver/AbstractClickReceiver.java
@@ -2,6 +2,7 @@
  * Apache 2.0 License
  *
  * Copyright (c) Sebastian Katzer 2017
+ * Contributor powowbox
  *
  * This file contains Original Code and/or Modifications of Original Code
  * as defined in and that are subject to the Apache License
@@ -21,7 +22,6 @@
 
 package de.appplant.cordova.plugin.notification.receiver;
 
-import android.app.IntentService;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
@@ -29,8 +29,6 @@ import android.os.Bundle;
 import de.appplant.cordova.plugin.notification.Manager;
 import de.appplant.cordova.plugin.notification.Notification;
 
-import static android.content.Intent.FLAG_ACTIVITY_REORDER_TO_FRONT;
-import static android.content.Intent.FLAG_ACTIVITY_SINGLE_TOP;
 import static de.appplant.cordova.plugin.notification.action.Action.CLICK_ACTION_ID;
 import static de.appplant.cordova.plugin.notification.action.Action.EXTRA_ID;
 
@@ -38,21 +36,22 @@ import static de.appplant.cordova.plugin.notification.action.Action.EXTRA_ID;
  * Abstract content receiver activity for local notifications. Creates the
  * local notification and calls the event functions for further proceeding.
  */
-abstract public class AbstractClickReceiver extends IntentService {
-
-    // Holds a reference to the intent to handle.
-    private Intent intent;
+abstract public class AbstractClickReceiver extends NotificationTrampolineActivity {
 
     public AbstractClickReceiver() {
-        super("LocalNotificationClickReceiver");
+      super();
+    }
+
+    public void onCreate(Bundle savedInstanceState) {
+      super.onCreate(savedInstanceState);
+      onHandleIntent(getIntent());
     }
 
     /**
      * Called when local notification was clicked to launch the main intent.
      */
-    @Override
     protected void onHandleIntent(Intent intent) {
-        this.intent        = intent;
+      // Holds a reference to the intent to handle.
 
         if (intent == null)
             return;
@@ -70,7 +69,6 @@ abstract public class AbstractClickReceiver extends IntentService {
             return;
 
         onClick(toast, bundle);
-        this.intent = null;
     }
 
     /**
@@ -86,12 +84,5 @@ abstract public class AbstractClickReceiver extends IntentService {
      */
     protected String getAction() {
         return getIntent().getExtras().getString(EXTRA_ID, CLICK_ACTION_ID);
-    }
-
-    /**
-     * Getter for the received intent.
-     */
-    protected Intent getIntent() {
-        return intent;
     }
 }