+++ src/ios/APPLocalNotification.m
@@ -2,6 +2,7 @@
  * Apache 2.0 License
  *
  * Copyright (c) Sebastian Katzer 2017
+ * Contributor Bhumin Bhandari
  *
  * This file contains Original Code and/or Modifications of Original Code
  * as defined in and that are subject to the Apache License
@@ -111,7 +112,9 @@ UNNotificationPresentationOptions const OptionAlert = UNNotificationPresentation
             oldNotification = [_center getNotificationWithId:id];
 
             if (oldNotification) {
-                [_center cancelNotification:oldNotification];
+                // Remove future alarms but keep ones that are active in the notification tray in case they need to persist.
+                // Call cancel() explicitly to keep the old behavior.
+                [self->_center removePendingNotificationRequestsWithIdentifiers:@[[id stringValue]]];
             }
 
             // Schedule the new notification
@@ -492,9 +495,12 @@ UNNotificationPresentationOptions const OptionAlert = UNNotificationPresentation
 {
     UNNotificationRequest* toast = notification.request;
 
-    [_delegate userNotificationCenter:center
+    if (_delegate != nil && ![toast.trigger isKindOfClass:UNCalendarNotificationTrigger.class]) {
+        [_delegate userNotificationCenter:center
               willPresentNotification:notification
                 withCompletionHandler:handler];
+        return;
+    }
 
     if ([toast.trigger isKindOfClass:UNPushNotificationTrigger.class])
         return;
@@ -524,9 +530,12 @@ UNNotificationPresentationOptions const OptionAlert = UNNotificationPresentation
 {
     UNNotificationRequest* toast = response.notification.request;
 
-    [_delegate userNotificationCenter:center
-       didReceiveNotificationResponse:response
-                withCompletionHandler:handler];
+    if (_delegate != nil && ![toast.trigger isKindOfClass:UNCalendarNotificationTrigger.class]) {
+        [_delegate userNotificationCenter:center
+           didReceiveNotificationResponse:response
+                    withCompletionHandler:handler];
+        return;
+    }
 
     handler();